Traditional_ML_Models/requirements.txt
scikit-learn>=1.0.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pillow>=9.0.0
streamlit>=1.28.0
tqdm>=4.62.0
pandas>=1.3.0
